# Files/Arquitetura-Java.openapi.json
{"openapi":"3.0.1","info":{"title":"Arquitetura-Java","description":"","version":"1.0.0"},"tags":[],"paths":{"/lojas":{"get":{"summary":"Lista Lojas","deprecated":false,"description":"","tags":[],"parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/lojas/buscar/{nome}":{"get":{"summary":"Busca Loja Por Nome","deprecated":false,"description":"","tags":[],"parameters":[{"name":"nome","in":"path","description":"","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/lojas/incluir":{"post":{"summary":"Adiciona Uma Nova Loja","deprecated":false,"description":"","tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"nome":{"type":"string"},"email":{"type":"string"},"endereco":{"type":"object","properties":{"cep":{"type":"string"}},"required":["cep"]}},"required":["nome","email","endereco"]}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/lojas/{id}/excluir":{"delete":{"summary":"Deleta uma Loja","deprecated":false,"description":"","tags":[],"parameters":[{"name":"id","in":"path","description":"","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/lojas/{id}":{"get":{"summary":"Busca Por ID","deprecated":false,"description":"","tags":[],"parameters":[{"name":"id","in":"path","description":"","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/veiculos/{id}":{"get":{"summary":"Obter Por Id","deprecated":false,"description":"","tags":[],"parameters":[{"name":"id","in":"path","description":"","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/veiculos/filtrarPorPreco":{"get":{"summary":"Filtro Por preço","deprecated":false,"description":"","tags":[],"parameters":[{"name":"min","in":"query","description":"","required":false,"schema":{"type":"number"}},{"name":"max","in":"query","description":"","required":false,"schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/veiculos/lista":{"get":{"summary":"Lista Todos","deprecated":false,"description":"","tags":[],"parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/veiculos/{id}/excluir":{"delete":{"summary":"Deletar","deprecated":false,"description":"","tags":[],"parameters":[{"name":"id","in":"path","description":"","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/veiculos/alterar":{"patch":{"summary":"Alterar Preço","deprecated":false,"description":"","tags":[],"parameters":[{"name":"id","in":"query","description":"","required":false,"schema":{"type":"integer"}},{"name":"preco","in":"query","description":"","required":false,"schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/carros/lista":{"get":{"summary":"Lista carros","deprecated":false,"description":"","tags":[],"parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/carros/incluir":{"post":{"summary":"Incluir carro","deprecated":false,"description":"","tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"modelo":{"type":"string"},"estoque":{"type":"boolean"},"preco":{"type":"number"},"portas":{"type":"integer"},"placa":{"type":"string"}},"required":["modelo","estoque","preco","portas","placa"]}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/carros/alterar":{"put":{"summary":"Alterar Carro","deprecated":false,"description":"","tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer"},"modelo":{"type":"string"},"estoque":{"type":"boolean"},"preco":{"type":"integer"},"portas":{"type":"integer"},"placa":{"type":"string"}},"required":["id","modelo","estoque","preco","portas","placa"]},"example":"    {\n        \"id\": 1,\n        \"modelo\": \"corolla\",\n        \"estoque\": true,\n        \"preco\": 70000.0,\n        \"portas\": 4,\n        \"placa\": \"BEE-3900\"\n    },"}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/bicicletas/lista":{"get":{"summary":"Lista bicicleta","deprecated":false,"description":"","tags":[],"parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/bicicletas/incluir":{"post":{"summary":"Incluir bicicleta","deprecated":false,"description":"","tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"modelo":{"type":"string"},"estoque":{"type":"boolean"},"preco":{"type":"number"},"marchas":{"type":"integer"},"freioDisco":{"type":"boolean"}},"required":["modelo","estoque","preco","marchas","freioDisco"]}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/bicicletas/alterar":{"put":{"summary":"Alterar bicicleta","deprecated":false,"description":"","tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer"},"modelo":{"type":"string"},"estoque":{"type":"boolean"},"preco":{"type":"integer"},"marchas":{"type":"integer"},"freioDisco":{"type":"boolean"}},"required":["id","modelo","estoque","preco","marchas","freioDisco"]},"example":" {\n        \"id\": 3,\n        \"modelo\": \"M8\",\n        \"estoque\": true,\n        \"preco\": 1000.0,\n        \"marchas\": 12,\n        \"freioDisco\": true\n    },"}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}}},"components":{"schemas":{},"securitySchemes":{}},"servers":[]}